Imp: 大きいファイルの暫定サポート
[patchunicode:#838]
Grepで2GBを超えるファイルも扱えるようにします
64bitビルドでは2GBを超えるファイルを開けるようにします


git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk2@3847 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CGrepAgent.h
@@ -137,7 +137,7 @@ private:
 		const TCHAR*	pszFilePath,	//	フルパス or 相対パス
 		const TCHAR*	pszCodeName,	//	文字コード情報"[SJIS]"とか
 		// マッチした行の情報
-		int				nLine,			//	マッチした行番号
+		LONGLONG		nLine,			//	マッチした行番号
 		int				nColumn,		//	マッチした桁番号
 		const wchar_t*	pCompareData,	//	行の文字列
 		int				nLineLen,		//	行の文字列の長さ
